+++ tests/Jroc.Tests/Node/Console/JavaScript/Require_NodeConsole_Log_Variadic.js
@@ -1,0 +1,4 @@
+"use strict";
+
+const nodeConsole = require("node:console");
+nodeConsole.log("values:", 1, 2, 3);